Moves useCart to separate hooks folder
