+++ src/components/ProductCard/index.jsx
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useCart } from "../../contexts/CartContext/useCart";
+import { useCart } from "../../hooks/useCart";
 import * as S from "./index.styled";
 import RenderStars from "../RenderStars";
