+++ src/hooks/useCart.js
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { CartContext } from "../CartContext";
+import { CartContext } from "../contexts/CartContext";
 
 export const useCart = () => useContext(CartContext);
